requirement.txt
black==23.9.1
blinker==1.6.2
click==8.1.7
colorama==0.4.6
flake8==6.1.0
Flask==2.3.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pycodestyle==2.11.0
pyflakes==3.1.0
Werkzeug==2.3.7
pymongo
flask
python-dotenv
requests
pymongo
